WIP: add basic example component, consolidate state layout while mimicing civic repo. GH-62
