+++ app/src/App.tsx
@@ -1,4 +1,5 @@
 import * as React from 'react';
+import Doohicky from './components/Doohicky'
 import './App.css';
 
 import logo from './logo.svg';
@@ -11,6 +12,7 @@ class App extends React.Component {
           <img src={logo} className="App-logo" alt="logo" />
           <h1 className="App-title">Welcome to React</h1>
         </header>
+        <Doohicky />
         <p className="App-intro">
           To get started, edit <code>src/App.tsx</code> and save to reload.
         </p>
@@ -18,5 +20,7 @@ class App extends React.Component {
     );
   }
 }
+
+
 
 export default App;
