+++ app/src/state/test/selectors.ts
@@ -2,7 +2,7 @@ import { createSelector } from 'reselect';
 
 
 export const rootState = (state: { test: any; }) => state.test || state;
-export const getCurrentState = createSelector(
+export const getCurrentState: any = createSelector(
   rootState,
   (state) => state
 );
