--- app/src/state/types/index.tsx
@@ -1,4 +1,0 @@
-// export interface StoreState {
-//   isLoading: boolean;
-//   error: boolean
-// }
